requirements.txt
# MCP Server Core
mcp>=1.2.0
mcp[cli]>=1.2.0
uvicorn>=0.23.0
fastapi>=0.104.0

# Media Processing
ffmpeg-python>=0.2.0
yt-dlp>=2023.11.14
Pillow>=10.0.0

# Database
supabase>=1.0.3
pinecone-client>=2.2.4
sqlalchemy>=2.0.23
alembic>=1.12.1

# AI and Vector Embeddings
openai>=1.3.0
numpy>=1.24.0
tenacity>=8.2.3
tiktoken>=0.5.1

# Caching and Performance
python-redis>=0.0.2
diskcache>=5.6.3

# Security and Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Utilities
pydantic>=2.4.2
python-dotenv>=1.0.0
loguru>=0.7.2
aiohttp>=3.8.6
httpx>=0.25.1
asyncio>=3.4.3
APScheduler>=3.10.4
aiofiles>=23.2.1
prometheus-client>=0.17.1
